Consolidate ParserConfig TypedDict profiles with inheritance preservation
