+++ src/datamodel_code_generator/_types/__init__.py
@@ -3,11 +3,13 @@
 from __future__ import annotations
 
 from datamodel_code_generator._types.generate_config_dict import GenerateConfigDict
-from datamodel_code_generator._types.graphql_parser_config_dict import GraphQLParserConfigDict
-from datamodel_code_generator._types.jsonschema_parser_config_dict import JSONSchemaParserConfigDict
-from datamodel_code_generator._types.openapi_parser_config_dict import OpenAPIParserConfigDict
 from datamodel_code_generator._types.parse_config_dict import ParseConfigDict
-from datamodel_code_generator._types.parser_config_dict import ParserConfigDict
+from datamodel_code_generator._types.parser_config_dicts import (
+    GraphQLParserConfig as GraphQLParserConfigDict,
+    JSONSchemaParserConfig as JSONSchemaParserConfigDict,
+    OpenAPIParserConfig as OpenAPIParserConfigDict,
+    ParserConfig as ParserConfigDict,
+)
 
 __all__ = [
     "GenerateConfigDict",
