+++ tox.ini
@@ -116,10 +116,7 @@ dependency_groups = pkg-meta
 description = Generate TypedDict files from config models (use --check to validate only)
 commands =
     datamodel-codegen --profile generate-config-dict {posargs}
-    datamodel-codegen --profile parser-config-dict {posargs}
-    datamodel-codegen --profile graphql-parser-config-dict {posargs}
-    datamodel-codegen --profile jsonschema-parser-config-dict {posargs}
-    datamodel-codegen --profile openapi-parser-config-dict {posargs}
+    datamodel-codegen --profile parser-config-dicts {posargs}
     datamodel-codegen --profile parse-config-dict {posargs}
 dependency_groups = dev
 no_default_groups = true
